+++ inc/defines.h
@@ -21,13 +21,33 @@
 
 // 127 and 126 on bash error codeja, cmd not found, ja cannot execute
 // bashissa 126 cannot-execute triggeroityy yleensa jos bad permissions tai shebang puuttuu
-# define CMD_ERROR_NOTFOUND		127
-# define CMD_ERROR_CANNOTEX		126
+# define ERROR_CMD_NOTFOUND		127
+# define ERROR_CMD_CANTEXEC		126
 # define ERROR					-1
 # define SUCCESS				0
 
-typedef struct s_state			t_state;
-typedef struct s_cmd			t_cmd;
-typedef enum e_cmd_type			t_cmd_type;
+typedef struct s_state
+{
+	int			pid_count;
+	pid_t		*pids;
+	int			exit_status; //edellisen exit status: tarvitaan $? komentoon
+	char		**env_var; //environment variableille
+	//whatever else is required to track globl status fo executions
+}	t_state;
+
+typedef struct s_cmd
+{
+	e_cmd_type		type;
+	char			**argv;
+	struct s_cmd	*left; //linked list kaytannossa.
+	struct s_cmd	*right;
+	int				exit_status;
+}	t_cmd;
+
+typedef enum e_cmd_type
+{
+	SIMPLE,
+	PIPELINE
+}	t_cmd_type;
 
 #endif
